requirements.txt
Django>=4.2,<6
django-import-export>=3.3
pandas>=2.0
plotly>=5.18
folium>=0.15.0
openpyxl>=3.0
gunicorn>=21.0
whitenoise>=6.0
pydeck>=0.8.0
# Opcionales: geopandas, altair, bokeh
